refactor, refetch every 10 sec when in viewport
