+++ package.json
@@ -68,6 +68,7 @@
     "react-dotdotdot": "^1.3.1",
     "react-dropzone": "^11.2.2",
     "react-helmet": "^6.1.0",
+    "react-intersection-observer": "^8.30.1",
     "react-markdown": "^5.0.2",
     "react-modal": "^3.11.2",
     "react-paginate": "^6.5.0",
